refactor: split delete link component
