--- app/feature/admin/components/types.ts
@@ -1,10 +1,0 @@
-export interface Link {
-  id: number;
-  url: string;
-  title: string;
-}
-
-export interface DialogProps {
-  isOpen: boolean;
-  onOpenChange: (open: boolean) => void;
-}
